Initial gamesim commit.
This is a toy workload generator, but does something kind of
interesting and provides a good base for doing more complex
things.  It simulates different users instantiating their state
by getting information from Membase, and then simulates
interactions between users, updating the state in Membase.

As of right now, this project is probably not much more than a
toy for building various demos.